# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Bichu API
  version: 1.0.0-rc.1
  summary: Identidade digital do pet, tag QR e rede local de perdido e achado.
  description: |
    Contrato unico entre o app Flutter, o front web que serve as paginas
    publicas (de outro time, desde o ADR-0017) e este backend.

    **Este servico nao renderiza HTML.** Toda resposta e JSON ou
    `application/problem+json`; as oito paginas publicas sairam daqui e sao
    montadas pelo front web a partir destas mesmas operacoes. Nao ha um segundo
    contrato para a web: o 410 com `next_action` chega la do mesmo jeito que
    chega ao app, e as duas pontas decidem pelo `type`, nunca pelo texto.

    ## Regra de acesso

    O app e navegavel deslogado; as **acoes** e que exigem conta. **Todas as 80
    operacoes declaram `security` explicitamente** — nenhuma herda o padrao da
    raiz, que existe so como rede de falha fechada. Rota sem `security`
    declarado nao sobe: a aplicacao derruba na carga do contrato.

    - `security: []` — publica, sem token.
    - `bearerAuth` — exige conta.
    - `bearerAuth` + `[]` — autenticacao **opcional**: a resposta muda se houver
      token (por exemplo, o dono da tag ve o modo dono).
    - `x-credential: tag_code_in_path` — **a posse do codigo impresso na tag e
      a credencial**, e ela viaja no caminho. Nao exige conta. E a excecao
      permanente: quem esta com a plaquinha na mao esta com o animal.

...
      para voltar a mesma conversa. Escopo de uma conversa so.

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []

# ==========================================================================
# Vocabularios fechados, em forma legivel por maquina
# ==========================================================================
#
# Existem porque um vocabulario que so vive em prosa (aqui, ou na 4.10 de
# `docs/04-seguranca.md`, ou no criterio de uma historia) tem tantas versoes
# quantos forem os lugares onde foi escrito, e elas divergem sem nenhum sinal.
# Foi o que aconteceu: a historia falava em 8 valores de `on_exceed`, a 4.10 em
# 9, e este contrato usa 11.
#
# A partir daqui o contrato e a fonte unica, e ele e verificavel: o lint le
# estas listas e reprova qualquer valor fora delas, em vez de carregar uma copia
# propria que envelhece. Documento que declare valores diferentes destes esta
# errado por definicao, e a correcao e no documento.
...
paths:
...
  /media/pet-photo-intents:
    post:
      operationId: createPetPhotoUploadIntent
      security:
        - bearerAuth: []
      tags: [pets]
      x-effects: [irreversible_write]
      summary: Pede as credenciais de upload direto para o armazenamento
      description: |
        O backend nunca recebe os bytes da imagem: ele assina a requisicao de
        upload para o armazenamento privado, com tipo e tamanho limitados.

        `method` diz a forma: `POST` com `fields` (politica de formulario, que
        e o caminho de hoje em MinIO e S3) ou `PUT` com `headers` assinados,
        para o provedor que nao suportar politica de POST. O cliente **le o
        campo** em vez de presumir a forma.
      x-rate-limit:
        - dimension: [account]
          limit: 20
          window: 1h
          on_exceed: deny_429
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/UploadIntentInput'
                - type: object
                  required: [pet_id]
                  properties:
                    pet_id: { type: string, format: uuid }
      responses:
        '201':
          description: Credenciais de upload, validas por 10 minutos.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UploadIntent' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '415': { $ref: '#/components/responses/UnsupportedMedia' }
...
  /found-reports:
    post:
      operationId: createStrayFoundReport
      security:
        - bearerAuth: []
      tags: [found]
      x-effects: [notifies, human_work, expensive_query]
      x-rate-limit:
        - dimension: [account]
          limit: 10
          window: 24h
          on_exceed: deny_429
        - dimension: [account]
          limit: 30
          window: 30d
          on_exceed: hold_for_review
      summary: Registra um achado avulso, sem QR
      description: |
        **Exige conta** (regra fechada no briefing). O cruzamento com casos
        abertos e por atributos e por distancia, e toda correspondencia passa
        por confirmacao humana do tutor.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StrayFoundReportInput' }
      responses:
        '201':
          description: Achado registrado.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FoundReport' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    get:
      operationId: listMyFoundReports
      security:
        - bearerAuth: []
      tags: [found]
      x-effects: []
      summary: Achados que eu registrei
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Lista.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FoundReportPage' }
        # BICHUS-35: `cursor` e `limit` PODEM recusar (maxLength, minimum,
        # maximum), entao a operacao precisa declarar o 400 que ela vai
        # responder. Sem esta linha `vigiarParametrosDasRotas` derruba a subida.
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /found-reports/{foundReportId}:
    parameters:
      - $ref: '#/components/parameters/FoundReportId'
    get:
      operationId: getFoundReport
      tags: [found]
      x-effects: []
      summary: Detalhe do achado
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Achado.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FoundReport' }
        # `foundReportId` e `format: uuid`, entao ele pode recusar e o 400
        # precisa estar declarado (BICHUS-35).
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      operationId: enrichFoundReport
      tags: [found]
      x-effects: [notifies]
      x-rate-limit:
        - dimension: [account]
          limit: 60
          window: 1h
          on_exceed: deny_429
      summary: Acrescenta detalhes ao aviso, para quem tem conta
      description: |
        Versao autenticada de `PATCH /v1/finder/found-report`, autorizada por
        **propriedade** (o aviso e de quem chama), nao por posse de token. Quem
        avisou sem conta usa a rota do achador, que nao tem id no caminho.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FoundReportEnrichment' }
      responses:
        '200':
          description: Aviso complementado.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FoundReport' }
        # Mesmo `foundReportId` de `getFoundReport`, mesma razao (BICHUS-35).
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '403': { $ref: '#/components/responses/Forbidden' }
        # O achado que nao e de quem chama responde 404, nunca 403 (ADR-0021):
        # um 403 confirmaria que aquele identificador e um achado de verdade.
        '404': { $ref: '#/components/responses/NotFound' }
        '410':
          description: Caso ja encerrado.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /media/found-report-photo-intents:
    post:
      operationId: createFoundReportPhotoUploadIntent
      tags: [found]
      x-effects: [irreversible_write]
      summary: Credenciais de upload da foto do achador
      description: |
        A foto do achador **nunca** aparece em rota publica: ela e vista pelo
        tutor dentro da conversa mediada, e so.
      security:
        - bearerAuth: []
      x-rate-limit:
        - dimension: [found_report]
          limit: 3
          window: lifetime
          on_exceed: deny_429
          note: "SEC-009: 3 fotos e 2 MB por foto no aviso do achador"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/UploadIntentInput'
                - type: object
                  required: [found_report_id]
                  properties:
                    found_report_id: { type: string, format: uuid }
      responses:
        '201':
          description: Credenciais de upload.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UploadIntent' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '415': { $ref: '#/components/responses/UnsupportedMedia' }

  # ----------------------------------------------------------------------
  # O achador sem conta: rotas enderecadas pelo token, sem id no caminho
  # ----------------------------------------------------------------------
  #
  # Existem separadas das rotas com id por decisao de seguranca (SEC-001): o par
  # "id no caminho + token no cabecalho" e o arranjo classico de BOLA, em que
  # basta a verificacao checar "token valido" em vez de "token pertence a este
  # objeto" para um achador ler a conversa de outro. Sem id no caminho, nao ha o
  # que adulterar, e o achador nunca recebe UUID interno nenhum.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Access token JWT assinado em RS256, 15 minutos de validade. `sub` e o
        UUID interno do usuario. A chave publica vem de
        `/.well-known/jwks.json`, selecionada pelo `kid` do cabecalho.
    reauth:
      type: apiKey
      in: header
      name: X-Reauth-Token
      description: |
        Segunda credencial das operacoes destrutivas, obtida em
        `POST /v1/auth/reauth`. Tres propriedades, e sem as tres os 5 minutos
        nao valem nada:

        - **uso unico**: consumido na primeira operacao que o apresenta;
        - **escopo por operacao**: emitido para **uma** finalidade declarada, e
          recusado nas demais. Sem isso, o tutor que reautentica para revogar uma
          plaquinha emite, sem saber, uma autorizacao para excluir a conta dele;
        - **vinculo com a sessao**: preso ao `jti` do token de acesso que o
          pediu, e invalidado por logout, troca de senha e
          `sessions_invalid_before`.

        A finalidade exigida por cada operacao esta em `x-reauth-scope`, porque
        requisito de seguranca de esquema `apiKey` nao aceita escopo na propria
        lista de `security`.
    webhookSignature:
      type: apiKey
      in: header
      name: X-Postmark-Signature
      description: |
        Assinatura do provedor, verificada com `MAIL_WEBHOOK_SECRET` em
        comparacao de tempo constante. Nao e credencial de usuario.
    finderToken:
      type: http
      scheme: bearer
      description: |
        Token opaco de 256 bits entregue a quem avisou o tutor sem ter conta.
...
  parameters:
...
    FoundReportId:
      name: foundReportId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Cursor:
      name: cursor
      in: query
      schema: { type: string, maxLength: 256 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Repetir a mesma chave em 24 h devolve a resposta original.
      schema: { type: string, format: uuid }
...
  responses:
...
    Forbidden:
      description: |
        Autenticado, mas sem permissao sobre este recurso. Para recurso de
        outro tutor a API responde **404**, nao 403: confirmar a existencia de
        um pet alheio ja e vazamento.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Nao existe, ou nao e visivel para quem perguntou.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    ValidationFailed:
      description: Corpo ou parametro invalido.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          examples:
            campoInvalido:
              value:
                type: https://dominio-a-definir.com.br/problems/validation-failed
                title: Dados invalidos
                status: 400
                detail: Um ou mais campos nao passaram na validacao.
                correlation_id: 018f7c1e-7a2b-7c3d-9e4f-2b1a0c9d8e7f
                errors:
                  - field: species
                    code: enum
                    message: Valor precisa ser dog, cat ou other.
...
    UnsupportedMedia:
      description: Tipo de arquivo nao aceito.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
...
    Species:
      type: string
      enum: [dog, cat, other]
    PetSize:
      type: string
      enum: [P, M, G, GG]
...
    UploadIntentInput:
      type: object
      required: [content_type, byte_size]
      properties:
        content_type:
          type: string
          enum: [image/jpeg, image/png, image/heic, image/webp]
        byte_size: { type: integer, minimum: 1, maximum: 10485760 }
    UploadIntent:
      type: object
      required: [upload_id, method, url, expires_at]
      properties:
        upload_id: { type: string, format: uuid }
        method:
          type: string
          enum: [POST, PUT]
          default: POST
          description: |
            A forma da requisicao de upload. Existe para que a troca do
            provedor de armazenamento seja configuracao: nem todo provedor
            compativel com S3 aceita politica de POST.
        url: { type: string, format: uri }
        fields:
          type: object
          additionalProperties: { type: string }
          description: Campos do formulario assinado. Presente quando `method` e POST.
        headers:
          type: object
          additionalProperties: { type: string }
          description: |
            Cabecalhos assinados que o cliente precisa enviar sem alterar
            (inclusive `Content-Type` e `Content-Length`). Presente quando
            `method` e PUT.
        expires_at: { type: string, format: date-time }
        max_bytes: { type: integer }
...
    StrayFoundReportInput:
      type: object
      description: |
        **Coordenada nao e obrigatoria**, pelo mesmo motivo do caso: quem negou
        a permissao informa `area`. Sem coordenada o cruzamento perde o
        criterio de distancia e passa a valer dentro da mesma cidade; a regra
        completa esta na secao 4.10 de `docs/03-arquitetura.md`.

        `share_token` existe para o vizinho que **chegou pelo push de um caso
        especifico** e esta dizendo "vi este pet": o achado e vinculado
        diretamente aquele caso, sem depender do cruzamento por atributos.
      required: [species, size, found_at]
      anyOf:
        - required: [location]
        - required: [area]
      properties:
        species: { $ref: '#/components/schemas/Species' }
        breed_code:
          type: string
          maxLength: 40
          description: |
            Mesma regra do pet: codigo de `reference-data`, e e ele que o
            cruzamento le. Quem acha um animal na rua raramente sabe a raca, e
            e por isso que este campo nao e obrigatorio aqui.
        breed_free_text:
          type: string
          maxLength: 40
          nullable: true
          description: |
            A raca como o achador escreveu. Descreve, nao cruza. So com
            `breed_code` em `outro_*`, pela mesma razao de `PetInput`.
        ref_data_version: { type: string, nullable: true }
        size: { $ref: '#/components/schemas/PetSize' }
        primary_color_code: { type: string, maxLength: 30 }
        sex: { $ref: '#/components/schemas/Sex' }
        found_at: { type: string, format: date-time }
        location: { $ref: '#/components/schemas/GeoPoint' }
        area: { $ref: '#/components/schemas/Area' }
        share_token:
          type: string
          description: |
            Opcional. Vincula o achado a um caso conhecido em vez de deixar o
            cruzamento descobrir. O candidato nasce com `status: suggested` e
            **continua exigindo confirmacao humana do tutor**.
        notes: { type: string, maxLength: 1000 }
        photo_upload_id: { type: string, format: uuid }
    FoundReport:
      type: object
      required: [id, origin, status, created_at]
      properties:
        id: { type: string, format: uuid }
        origin: { type: string, enum: [tag_scan, stray_report] }
        status: { type: string, enum: [open, matched, closed] }
        species: { $ref: '#/components/schemas/Species' }
        size: { $ref: '#/components/schemas/PetSize' }
        area_label:
          type: string
          nullable: true
          description: Bairro. O ponto exato nunca sai desta forma.
        found_at: { type: string, format: date-time }
        photo_url: { type: string, format: uri, nullable: true }
        notes: { type: string, nullable: true }
        conversation_id: { type: string, format: uuid, nullable: true }
        created_at: { type: string, format: date-time }
...
    FoundReportPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/FoundReport' }
        next_cursor: { type: string, nullable: true }
...